+++ R/predictions.R
@@ -1,6 +1,6 @@
 # select prediction method, based on model-object
 #' @importFrom sjstats link_inverse
-select_prediction_method <- function(fun, model, expanded_frame, ci.lvl, type, faminfo, ppd, terms, typical, ...) {
+select_prediction_method <- function(fun, model, expanded_frame, ci.lvl, type, faminfo, ppd, terms, typical, prettify, ...) {
   # get link-inverse-function
   linv <- sjstats::link_inverse(model)
 
@@ -27,7 +27,7 @@ select_prediction_method <- function(fun, model, expanded_frame, ci.lvl, type, f
     fitfram <- get_predictions_glmmTMB(model, expanded_frame, ci.lvl, linv, ...)
   } else if (fun %in% c("lmer", "nlmer", "glmer")) {
     # merMod-objects  -----
-    fitfram <- get_predictions_merMod(model, expanded_frame, ci.lvl, linv, type, terms, typical, ...)
+    fitfram <- get_predictions_merMod(model, expanded_frame, ci.lvl, linv, type, terms, typical, prettify, ...)
   } else if (fun == "gam") {
     # gam-objects -----
     fitfram <- get_predictions_gam(model, expanded_frame, ci.lvl, linv, ...)
@@ -36,7 +36,7 @@ select_prediction_method <- function(fun, model, expanded_frame, ci.lvl, type, f
     fitfram <- get_predictions_vgam(model, expanded_frame, ci.lvl, linv, ...)
   } else if (fun %in% c("lme", "gls", "plm")) {
     # lme-objects -----
-    fitfram <- get_predictions_lme(model, expanded_frame, ci.lvl, terms, typical, ...)
+    fitfram <- get_predictions_lme(model, expanded_frame, ci.lvl, terms, typical, prettify, ...)
   } else if (fun == "gee") {
     # gee-objects -----
     fitfram <- get_predictions_gee(model, expanded_frame, linv, ...)
@@ -54,7 +54,7 @@ select_prediction_method <- function(fun, model, expanded_frame, ci.lvl, type, f
     fitfram <- get_predictions_polr(model, expanded_frame, linv, ...)
   } else if (fun %in% c("betareg", "truncreg", "zeroinfl", "hurdle")) {
     # betareg, truncreg, zeroinfl and hurdle-objects -----
-    fitfram <- get_predictions_generic2(model, expanded_frame, ci.lvl, fun, typical, terms, ...)
+    fitfram <- get_predictions_generic2(model, expanded_frame, ci.lvl, fun, typical, terms, prettify, ...)
   } else if (fun %in% c("glm", "glm.nb")) {
     # glm-objects -----
     fitfram <- get_predictions_glm(model, expanded_frame, ci.lvl, linv, ...)
@@ -289,7 +289,7 @@ get_predictions_clm <- function(model, fitfram, ci.lvl, linv, ...) {
 
 # predictions for regression models w/o SE ----
 
-get_predictions_generic2 <- function(model, fitfram, ci.lvl, fun, typical, terms, ...) {
+get_predictions_generic2 <- function(model, fitfram, ci.lvl, fun, typical, terms, prettify, ...) {
   # get prediction type.
   pt <- dplyr::case_when(
     fun %in% c("hurdle", "zeroinfl") ~ "response",
@@ -320,7 +320,8 @@ get_predictions_generic2 <- function(model, fitfram, ci.lvl, fun, typical, terms
       fitfram = fitfram,
       typical = typical,
       terms = terms,
-      fun = fun
+      fun = fun,
+      prettify = prettify
     )
 
   se.fit <- se.pred$se.fit
@@ -441,7 +442,7 @@ get_predictions_glmmTMB <- function(model, fitfram, ci.lvl, linv, ...) {
 
 # predictions for merMod ----
 
-get_predictions_merMod <- function(model, fitfram, ci.lvl, linv, type, terms, typical, ...) {
+get_predictions_merMod <- function(model, fitfram, ci.lvl, linv, type, terms, typical, prettify, ...) {
   # does user want standard errors?
   se <- !is.null(ci.lvl) && !is.na(ci.lvl)
 
@@ -475,7 +476,8 @@ get_predictions_merMod <- function(model, fitfram, ci.lvl, linv, type, terms, ty
         fitfram = fitfram,
         typical = typical,
         terms = terms,
-        type = type
+        type = type,
+        prettify = prettify
       )
 
     se.fit <- se.pred$se.fit
@@ -789,7 +791,7 @@ get_predictions_lm <- function(model, fitfram, ci.lvl, linv, ...) {
 #' @importFrom sjstats resp_var pred_vars
 #' @importFrom purrr map
 #' @importFrom tibble add_column
-get_predictions_lme <- function(model, fitfram, ci.lvl, terms, typical, ...) {
+get_predictions_lme <- function(model, fitfram, ci.lvl, terms, typical, prettify, ...) {
   # does user want standard errors?
   se <- !is.null(ci.lvl) && !is.na(ci.lvl)
 
@@ -813,7 +815,7 @@ get_predictions_lme <- function(model, fitfram, ci.lvl, terms, typical, ...) {
 
   # did user request standard errors? if yes, compute CI
   if (se) {
-    se.pred <- get_se_from_vcov(model = model, fitfram = fitfram, typical = typical, terms = terms)
+    se.pred <- get_se_from_vcov(model = model, fitfram = fitfram, typical = typical, terms = terms, prettify = prettify)
 
     se.fit <- se.pred$se.fit
     fitfram <- se.pred$fitfram
@@ -944,7 +946,7 @@ get_base_fitfram <- function(fitfram, linv, prdat, se, ci.lvl) {
 #' @importFrom dplyr arrange
 #' @importFrom sjstats resp_var model_frame
 #' @importFrom rlang parse_expr
-get_se_from_vcov <- function(model, fitfram, typical, terms, fun = NULL, type = "fe") {
+get_se_from_vcov <- function(model, fitfram, typical, terms, fun = NULL, type = "fe", prettify = TRUE) {
   # copy data frame with predictions
   newdata <- get_expanded_data(
     model,
@@ -952,7 +954,8 @@ get_se_from_vcov <- function(model, fitfram, typical, terms, fun = NULL, type = 
     terms,
     typ.fun = typical,
     fac.typical = FALSE,
-    type = type
+    type = type,
+    prettify = prettify
   )
 
   # add response